+++ src/Controller/CheckInput.java
@@ -1,12 +1,14 @@
+package Controller;
+
 import java.util.Scanner;
 
 /** 
  * Static functions used to check console input for validity.
  *
- * Use:	Place CheckInput class in the same project folder as your code.
- *	Call CheckInput functions from your code using "CheckInput."
+ * Use:	Place Controller.CheckInput class in the same project folder as your code.
+ *	Call Controller.CheckInput functions from your code using "Controller.CheckInput."
  *
- * Example:  int num = CheckInput.getInt();
+ * Example:  int num = Controller.CheckInput.getInt();
  *
  * @author Shannon Cleary 2021
  */
@@ -129,51 +131,5 @@ public class CheckInput {
 			}
 		}
 		return false;
-	}
-
-	/**
-	 * Takes in a string that has a dollar sign.
-	 * Removes the $ symbol and adds the value into a double
-	 * @return the double without the dollar sign.
-	 * **/
-	public static double checkDollarSign() {
-		Scanner in = new Scanner(System.in);
-		boolean isRunning = true;
-		while (isRunning) {
-			if (in.hasNextDouble()) {
-				System.out.println("the user inputted a double");
-				double amount = in.nextDouble();
-				isRunning = false;
-				return amount;
-			}
-			if (in.hasNextLine()) {
-				String input = in.nextLine();
-				char[] stringToInt = new char[input.length()];
-				if (input.charAt(0) == '$') {
-					for (int i = 0; i < input.length() - 1; i++) {
-						stringToInt[i] = input.charAt(i + 1);
-					}
-					try {
-						String s = String.valueOf(stringToInt);
-						double amount = Double.parseDouble(s);
-						isRunning = false;
-						return amount;
-					} catch (Exception e) {
-						System.out.println("\n-----------------------------------------------------------------------------");
-						System.out.println("Invalid Input");
-						System.out.println("Please input with the correct format: [ $(amount) ]");
-						System.out.println("-----------------------------------------------------------------------------");
-						System.out.print("Amount: ");
-					}
-				} else {
-					System.out.println("\n-----------------------------------------------------------------------------");
-					System.out.println("Invalid Input");
-					System.out.println("Please input with the correct format: [ $(amount) ]");
-					System.out.println("-----------------------------------------------------------------------------");
-					System.out.print("Amount: ");
-				}
-			}
-		}
-		return 0;
 	}
 }
