--- src/Fire.java
@@ -1,8 +1,0 @@
-public interface Fire {
-    String specialMenu = "1. Ember\n2. Fire Blast\n3. Fire Punch";
-    int numSpecialMenuItems = 3;
-
-    String ember(Pokemon p);
-    String fireBlast(Pokemon p);
-    String firePunk(Pokemon p);
-}
