--- src/Pokemon.java
@@ -1,6 +1,0 @@
-public abstract class Pokemon extends Entity{
-    public Pokemon(String n){
-        super(n, 100);
-    }
-
-}
