--- src/Staryu.java
@@ -1,16 +1,0 @@
-
-public class Staryu extends Pokemon implements Water{
-	public Staryu(String n) {
-		super(n);
-	}
-
-	public String watergun(Pokemon p) {
-		return "Water Gun attack!";
-	}
-	public String bubblebeam(Pokemon p) {
-		return "Bubble Beam attack!";
-	}
-	public String waterfall(Pokemon p) {
-		return "Waterfall attack!";
-	}
-}
